Move connections
