+++ src/cow-react/modules/wallet/web3-react/connection/NetworkOption.tsx
@@ -5,7 +5,7 @@ import { Network } from '@web3-react/network'
 
 import { RPC_URLS } from 'constants/networks'
 import { ConnectionType } from '@cow/modules/wallet'
-import { Connection } from '../../utils/connection'
+import { Connection } from '../utils/connection'
 
 
 const [web3Network, web3NetworkHooks] = initializeConnector<Network>(
