+++ src/custom/utils/switchChain.ts
@@ -1,9 +1,9 @@
-import { coinbaseWalletConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/CoinbaseWalletOption'
-import { fortmaticConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/FortmaticOption'
-import { injectedConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/InjectedOption'
-import { networkConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/NetworkOption'
-import { gnosisSafeConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/SafeOption'
-import { walletConnectConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/WalletConnectOption'
+import { coinbaseWalletConnection } from '@cow/modules/wallet/web3-react/connection/CoinbaseWalletOption'
+import { fortmaticConnection } from '@cow/modules/wallet/web3-react/connection/FortmaticOption'
+import { injectedConnection } from '@cow/modules/wallet/web3-react/connection/InjectedOption'
+import { networkConnection } from '@cow/modules/wallet/web3-react/connection/NetworkOption'
+import { gnosisSafeConnection } from '@cow/modules/wallet/web3-react/connection/SafeOption'
+import { walletConnectConnection } from '@cow/modules/wallet/web3-react/connection/WalletConnectOption'
 import { Connector } from '@web3-react/types'
 import { getChainInfo } from 'constants/chainInfo'
 import { ALL_SUPPORTED_CHAIN_IDS, SupportedChainId } from 'constants/chains'
