+++ src/utils/switchChain.ts
@@ -1,10 +1,10 @@
 import { Connector } from '@web3-react/types'
-import { coinbaseWalletConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/CoinbaseWalletOption'
-import { fortmaticConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/FortmaticOption'
-import { injectedConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/InjectedOption'
-import { networkConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/NetworkOption'
-import { gnosisSafeConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/SafeOption'
-import { walletConnectConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/WalletConnectOption'
+import { coinbaseWalletConnection } from '@cow/modules/wallet/web3-react/connection/CoinbaseWalletOption'
+import { fortmaticConnection } from '@cow/modules/wallet/web3-react/connection/FortmaticOption'
+import { injectedConnection } from '@cow/modules/wallet/web3-react/connection/InjectedOption'
+import { networkConnection } from '@cow/modules/wallet/web3-react/connection/NetworkOption'
+import { gnosisSafeConnection } from '@cow/modules/wallet/web3-react/connection/SafeOption'
+import { walletConnectConnection } from '@cow/modules/wallet/web3-react/connection/WalletConnectOption'
 import { getChainInfo } from '@src/constants/chainInfo'
 import { ALL_SUPPORTED_CHAIN_IDS, SupportedChainId } from '@src/constants/chains'
 import { RPC_URLS } from '@src/constants/networks'
